first commit, starting foundry playground

+++ src/Counter.sol
@@ -2,10 +2,10 @@
 pragma solidity ^0.8.21;
 
 contract Counter {
-    uint public count;
+    uint256 public count;
 
     // Function to get the current count
-    function get() public view returns (uint) {
+    function get() public view returns (uint256) {
         return count;
     }
